Replace insert_number with insert_numbers to support inserting multiple entries
This commit is quite satisfying because it simultaneously simplifies the
algorithm used, makes it more capable (can insert more than one entry) and also
improves the quality of solutions given (it tests a sliding window of
renamings, not just "all before" or "all afterwards").

There is a notable regression which is that we no longer provide a series of
renamings which guarantee freedom from collisions. Therefore all renamed files
must first be moved to a temporary location then moved back into their new
locations to avoid accidentally overwriting things.